+++ packages/primitives/src/components/Radio/Radio.types.ts
@@ -1,0 +1,20 @@
+import { ChangeEvent, ReactNode } from 'react';
+
+export interface RadioState {
+  selectedValue: string;
+}
+
+export interface RadioProps extends RadioState {
+  name: string;
+  children: ReactNode;
+  onChange?: (e: ChangeEvent<HTMLInputElement>) => void;
+}
+
+export type RadioOptionProps = {
+  id?: string;
+  name?: string;
+  value: string;
+  children: ReactNode;
+  disabled?: boolean;
+  onChange?: (e: ChangeEvent<HTMLInputElement>) => void;
+};
