+++ packages/primitives/src/components/Radio/index.ts
@@ -1,0 +1,9 @@
+import RadioRoot from './radio';
+import Option from './RadioOption';
+
+const Radio = Object.assign(RadioRoot, {
+  Option,
+});
+
+export { Radio };
+export * from './Radio.types';
